requirements.txt
discord.py
aiohttp
pandas
matplotlib
python-dotenv